Created Color converter and PPM converter to map data from those objects to and from their respective DTOs

+++ Controllers/DTOs/ColorDTO.cs
@@ -8,9 +8,9 @@ namespace Controllers.DTOs
 {
     public class ColorDTO
     {
-        public int Red { get; set; }
-        public int Green { get; set; }
-        public int Blue { get; set; }
+        public double Red { get; set; }
+        public double Green { get; set; }
+        public double Blue { get; set; }
 
         public ColorDTO() { }
